requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
patchright>=1.0.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
requests>=2.31.0
PyJWT>=2.8.0
